+++ src/architectures/kind.ts
@@ -1,6 +1,7 @@
 export enum Kind {
   arm64,
-  x86_64
+  x86_64,
+  riscv64
 }
 
 export function toKind(value: string): Kind | undefined {
@@ -12,5 +13,8 @@ const architectureMap: Record<string, Kind> = {
   aarch64: Kind.arm64,
   'x86-64': Kind.x86_64,
   x86_64: Kind.x86_64,
-  x64: Kind.x86_64
+  x64: Kind.x86_64,
+  riscv64: Kind.riscv64,
+  riscv: Kind.riscv64,
+  rv64: Kind.riscv64
 } as const