+++ changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add support for FreeBSD on RISC-V 64 (`riscv64`)
+
 ### Fixed
 - Don't log the `Tearing down VM` group when `shutdown_vm` is `false`, since
     the VM is not being torn down in that case